requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
Flask==2.3.3
gunicorn==21.2.0